% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multired.R
\name{compute_JSD_matrix}
\alias{compute_JSD_matrix}
\title{Computes Jensen-Shannon matrix}
\usage{
compute_JSD_matrix(m, verbose = FALSE)
}
\arguments{
\item{m}{multilayer network}

\item{verbose}{if TRUE, outputs progress text and bars}
}
\description{
Computes Jensen-Shannon matrix
}

