% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multired.R
\name{get_adjacency_of_layer}
\alias{get_adjacency_of_layer}
\title{Adjacency matrix of one or more layers}
\usage{
get_adjacency_of_layer(layerId, m)
}
\arguments{
\item{layerId}{An integer identifying one layer of the network.}

\item{m}{A multilayer object}
}
\value{
A \code{\link[Matrix]{Matrix}} object
}
\description{
The function will return the adjacency matrix of one layer of the multilayer network
}

