% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multired.R
\name{get_agg_entropy}
\alias{get_agg_entropy}
\title{Entropy of the aggregation of all layers}
\usage{
get_agg_entropy(m)
}
\arguments{
\item{m}{multilayer network}
}
\description{
Entropy of the aggregation of all layers
}

