% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multired.R
\name{get_entropy}
\alias{get_entropy}
\title{Calculates entropy of multilayer network}
\usage{
get_entropy(m)
}
\arguments{
\item{m}{multilayer network}
}
\description{
Calculates entropy of multilayer network
}

