% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multired.R
\name{get_layer_entropy}
\alias{get_layer_entropy}
\title{Adjacency matrix of one or more layers}
\usage{
get_layer_entropy(A)
}
\arguments{
\item{A}{adjacency matrix.}
}
\value{
The entropy of the adjacency matrix.
}
\description{
The function will return the adjacency matrix of one or more layers of the multilayer network,
 if more than one layerID is passed, the function will aggregate the layers,
 summing the weights of the links.
}

