% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multired.R
\name{get_relative_entropy}
\alias{get_relative_entropy}
\title{Calculates relative entropy (q) for reduced multilayer network}
\usage{
get_relative_entropy(reducedM, aggEntropy)
}
\arguments{
\item{reducedM}{the multilayer network with reduced number of layers}

\item{aggEntropy}{the aggregated entropy of the original multilayer. See \code{\link{get_agg_entropy}}.}
}
\value{
The relative entropy (q)
}
\description{
Calculates the relative entropy of a reduced multilayer, 
relative to the aggregated entropy of the original multilayer (not the reduced).
}

