% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multired.R
\name{list_to_multilayer}
\alias{list_to_multilayer}
\title{Converts a List of Matrix to Data frame}
\usage{
list_to_multilayer(l, n, networkName = NULL)
}
\description{
Each position in the list must be a n x n matrix
}

