% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multired.R
\name{read_multilayer}
\alias{read_multilayer}
\title{Read a multilayer network from a text file}
\usage{
read_multilayer(filename, n, networkName = NULL)
}
\arguments{
\item{filename}{The path to the file}

\item{n}{The total number of nodes in the multilayer network.
If not given, multired will infer the number of nodes from the edge list automatically,
but depending on the size of the network, this might take a while.}

\item{networkName}{The name of the network (optional).}
}
\value{
A multilayer object
}
\description{
The read.multilayer function reads only text files.
}
\details{
The file must contain three or four columns, each line with the following format:

<layerID> <source> <target> (<weight>)

If the fourth column is not informed, all edges will be assigned weight 1.
}

