% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multired.R
\name{get_adjacency_of_layer}
\alias{get_adjacency_of_layer}
\title{Adjacency matrix of one or more layers}
\usage{
get_adjacency_of_layer(layerIds, m)
}
\arguments{
\item{layerIds}{An integer or a vector, identifying one or more layerID of the network.
If layerIds is a vector, the layers will be aggregated.}

\item{m}{A multilayer object}
}
\value{
A \code{\link[Matrix]{Matrix}} object
}
\description{
The function will return the adjacency matrix of one or more layers of the multilayer network,
 if more than one layerID is passed, the function will aggregate the layers,
 summing the weights of the links.
}

