% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multired.R
\name{get_layer_entropy}
\alias{get_layer_entropy}
\title{Entropy of an adjacency matrix}
\usage{
get_layer_entropy(A)
}
\arguments{
\item{A}{adjacency matrix.}
}
\value{
The entropy of the adjacency matrix.
}
\description{
Entropy of an adjacency matrix
}

